Move BigUint serialization to a library.
