+++ crates/cairo-lang-utils/Cargo.toml
@@ -10,11 +10,15 @@ description = "General utilities for the Cairo compiler project."
 testing = []
 
 [dependencies]
+
 chrono.workspace = true
+serde.workspace = true
 env_logger.workspace = true
 indexmap.workspace = true
 itertools.workspace = true
 log.workspace = true
+num-bigint.workspace = true
+num-traits.workspace = true
 
 [dev-dependencies]
 test-case.workspace = true
