+++ crates/cairo-lang-utils/src/bigint.rs
@@ -1,0 +1,39 @@
+use num_bigint::BigUint;
+use num_traits::Num;
+use serde::ser::Serializer;
+use serde::{Deserialize, Deserializer, Serialize};
+
+// A wrapper for BigUint that serializes as hex.
+#[derive(Default, Debug, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(transparent)]
+pub struct BigUintAsHex {
+    /// A field element that encodes the signature of the called function.
+    #[serde(serialize_with = "serialize_big_uint", deserialize_with = "deserialize_big_uint")]
+    pub value: BigUint,
+}
+
+fn deserialize_from_str<'a, D>(s: &str) -> Result<BigUint, D::Error>
+where
+    D: Deserializer<'a>,
+{
+    match s.strip_prefix("0x") {
+        Some(num_no_prefix) => BigUint::from_str_radix(num_no_prefix, 16)
+            .map_err(|error| serde::de::Error::custom(format!("{error}"))),
+        None => Err(serde::de::Error::custom(format!("{s} does not start with `0x` is missing."))),
+    }
+}
+
+pub fn serialize_big_uint<S>(num: &BigUint, serializer: S) -> Result<S::Ok, S::Error>
+where
+    S: Serializer,
+{
+    serializer.serialize_str(&format!("{num:#x}"))
+}
+
+pub fn deserialize_big_uint<'a, D>(deserializer: D) -> Result<BigUint, D::Error>
+where
+    D: Deserializer<'a>,
+{
+    let s = &String::deserialize(deserializer)?;
+    deserialize_from_str::<D>(s)
+}
